DHS-1601: add retries to connect to LDAP even after initial connection
This is a first draft of this change.

We don't want to restart the container automatically. We want to know
when something goes wrong that we didn't expect. But we expect LDAP
server to be down every now and again. So far, our re-tries limited
themselves to the initial establishing of the connection.

We also except that the LDAP server might go down momentarily midway
through the execution of the sram-sync, which is what this commit
implements.