Fix re-try mechanism for establishing iRODS session
Simply instantiating iRODSSession() won't actually attempt to talk to
the server.To force it to actually establish connection, querying the
server's iRODS version seems to do the trick.

Ideally we would be sure that the entirety of sram-sync is resistant to
temporary loss of upstream services. And it does _look_ like it might be
safe to just restart from the top, even if we crashed (dropped
connection to iRODS) midway.

But I'm not 100% sure at the moment that both our code (sram-sync) and
python-irodsclient itself have this property (safe to crash midway
anywhere & restart). So, for now, re-try mechanism for iRODS only happens
during the initial establishing of the connection. If iRODS dies
sometime after, sram-sync will crash.

With LDAP, our re-try mechanism will get triggered even if LDAP dies
midway of some transition, hopefully this is fine but I have not gone
through the code to verify that there is no possibility of ending up in
a mangled state.

[DHDO-763]