Add PoC dh_is_ready.sh
In a rudimentary way, we simply have sram-sync write down how many runs
has it gone through in a file, that then we can read from
dh_is_ready.sh. We will use this later to wait on it in development.

In the future maybe we could have it write to that file something more
meaningful..

[DHDO-765]